# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Makers Community API",
        "description": "API complète pour la plateforme communautaire Makers Community avec authentification, posts, commentaires, ratings et modération automatique. Cette API permet de gérer une communauté de makers avec toutes les fonctionnalités nécessaires.",
        "contact": {
            "name": "Équipe Makers Community",
            "email": "dev@makerscomm.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.makerscomm.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/posts/{post}/report": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Signaler un post",
                "description": "Signaler un post pour contenu inapproprié. Déclenche des actions automatiques selon le nombre de signalements (3 = avertissement, 5 = désactivation).",
                "operationId": "c136185649c7b8302d47d3497a0bc19a",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID du post à signaler",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "description": "Raison du signalement",
                                        "type": "string",
                                        "enum": [
                                            "spam",
                                            "inappropriate",
                                            "harassment",
                                            "fake",
                                            "other"
                                        ],
                                        "example": "spam"
                                    },
                                    "description": {
                                        "description": "Description détaillée (optionnel, max 500 caractères)",
                                        "type": "string",
                                        "example": "Ce post contient du spam publicitaire",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Signalement créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "error": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "properties": {
                                                "report": {
                                                    "$ref": "#/components/schemas/Report"
                                                },
                                                "total_reports": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Post signalé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Déjà signalé ou erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Vous avez déjà signalé ce post"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Signalement déjà effectué"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/{post}/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Lister les signalements d'un post",
                "description": "Récupérer tous les signalements d'un post avec statistiques par raison",
                "operationId": "e181ea3490543cfe470682879e1bb3b9",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signalements récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "error": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "properties": {
                                                "post_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "reports": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Report"
                                                    }
                                                },
                                                "total_reports": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "reports_by_reason": {
                                                    "properties": {
                                                        "spam": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "inappropriate": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "harassment": {
                                                            "type": "integer",
                                                            "example": 0
                                                        },
                                                        "fake": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "other": {
                                                            "type": "integer",
                                                            "example": 0
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Signalements récupérés avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/reports/{report}": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Traiter un signalement",
                "description": "Marquer un signalement comme traité avec une note administrative (Admin uniquement)",
                "operationId": "e168316ca64c3b54277bf430869a47ba",
                "parameters": [
                    {
                        "name": "report",
                        "in": "path",
                        "description": "ID du signalement",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Nouveau statut du signalement",
                                        "type": "string",
                                        "enum": [
                                            "reviewed",
                                            "resolved",
                                            "rejected"
                                        ],
                                        "example": "resolved"
                                    },
                                    "admin_note": {
                                        "description": "Note administrative (optionnel, max 500 caractères)",
                                        "type": "string",
                                        "example": "Contenu supprimé après vérification",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Signalement traité avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "error": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "properties": {
                                                "report": {
                                                    "$ref": "#/components/schemas/Report"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Signalement traité avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Signalement non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Report": {
                "title": "Signalement",
                "description": "Signalement de contenu",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "reason": {
                        "type": "string",
                        "enum": [
                            "spam",
                            "inappropriate",
                            "harassment",
                            "fake",
                            "other"
                        ],
                        "example": "spam"
                    },
                    "description": {
                        "type": "string",
                        "example": "Ce post contient du spam publicitaire",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "reviewed",
                            "resolved",
                            "rejected"
                        ],
                        "example": "pending"
                    },
                    "reporter_user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "resolved_by": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "admin_note": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "resolved_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T17:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T17:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "StandardResponse": {
                "title": "Réponse Standard",
                "description": "Format de réponse standardisé",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "success",
                            "error"
                        ],
                        "example": "success"
                    },
                    "error": {
                        "example": null,
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "type": "string"
                            },
                            {
                                "type": "object"
                            }
                        ]
                    },
                    "data": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "type": "object"
                            },
                            {
                                "type": "array",
                                "items": {}
                            }
                        ]
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Erreur de Validation",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "error": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Ce champ est requis."
                                }
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "null"
                    },
                    "message": {
                        "type": "string",
                        "example": "Erreur de validation"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token d'authentification Laravel Sanctum. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}